NEW SALES INVENTORY WINDOW
It seems that I misunderstood our instructor's instructions, so i
decided to update the way the sales inventory is displayed, instead of a
JOptionPane, I decided to create a separate JFrame for it, It is not
really necessary but I think it just makes it more better-looking.

One of the most frustrating parts was changing how the sales inventories
are stored

i had to compare two singleton classes, the Order and the
SalesInventory,

the SalesInventory class contains a new BoughtItem[]

so everytime the user completes an order, we already know that the
their order gets saved in the ArrayList in the Order class right?
Then, using nested for loops, for every BoughtItem object in the Order
ArrayList, it will get "matched" with a particular object in the array inside
the SalesInventory class, if it matches, then the particular BoughtItem
in the SalesInventory array has its quantity appended, depending
on the quantity of the particular BoughtItem object in the Order
ArrayList